update my team
