+++ controllers/teamController.js
@@ -3,9 +3,10 @@ const SerTeam = require('../services/serTeam')
 const getTeamList = async (req, res) => {
   const serTeam = new SerTeam()
   const teamList = await serTeam.findAllTeamListWithoutMe(1)// token
-  res.json({
+  res.status(200).json({
     data: teamList
   })
+  //res 401: Unauthorized
 }
 
 // 팀 생성
@@ -38,38 +39,9 @@ const createTeam = async (req, res, next) => {
   }
 }
 
-// 팀 수정
-const updateMyTeam = async (req, res) => {
-  const serTeam = new SerTeam()
-  const id = 5
-  const {
-    body: {
-      name,
-      chat_address,
-      owner_id,
-      intro,
-      password,
-      max_member_number
-    }
-  } = req
-  try {
-    await serTeam.updateMyTeam({
-      id,
-      name,
-      chat_address,
-      owner_id,
-      intro,
-      password,
-      max_member_number
-    })
-    res.status(202).json('팀 수정 성공')
-  } catch (error) {
-    console.log(error)
-  }
-}
+
 
 module.exports = {
   getTeamList,
   createTeam,
-  updateMyTeam
 }
