+++ routes/v1/api/teamRouter.js
@@ -1,10 +1,9 @@
 const express = require('express')
-const { getTeamList,createTeam,updateMyTeam } = require('../../../controllers/teamController')
+const { getTeamList,createTeam } = require('../../../controllers/teamController')
 
 const router = express.Router()
 
 router.get('/',getTeamList)
 router.post('/', createTeam)
-router.patch('/', updateMyTeam)
 
 module.exports = router
