+++ services/serMe.js
@@ -3,13 +3,13 @@ const ModelTeam = require('../models/modTeam')
 const ModelBelong = require('../models/modBelong')
 
 class serMe {
-  constructor () {
+  constructor() {
     this.modUser = new ModelUser()
     this.modTeam = new ModelTeam()
     this.modBelong = new ModelBelong()
   }
 
-  async findMyInfo (userId) {
+  async findMyInfo(userId) {
     try {
       const myInfo = await this.modUser.findUserInfoById(userId)
       return myInfo
@@ -18,7 +18,7 @@ class serMe {
     }
   }
 
-  async findMyTeamList (userId) {
+  async findMyTeamList(userId) {
     try {
       const teamListOwner = await this.modTeam.findMyTeamList(userId).then()
       const teamListId = await this.modBelong.findMyTeamList(userId).then()
@@ -29,12 +29,21 @@ class serMe {
     }
   }
 
-  async updateMyInfo (data) {
+  async updateMyInfo(data) {
     try {
       await this.modUser.updateUserInfo(data)
     } catch (error) {
       console.log(error)
     }
   }
+
+  async updateMyTeam (data) {
+    try {
+      await this.modTeam.updateUserTeam(data)
+    } catch (error) {
+      console.log(error)
+    }
+  }
+
 }
 module.exports = serMe
