+++ services/serTeam.js
@@ -27,12 +27,5 @@ class SerTeam {
     }
   }
 
-  async updateMyTeam (data) {
-    try {
-      await this.modTeam.updateUserTeam(data)
-    } catch (error) {
-      console.log(error)
-    }
-  }
 }
 module.exports = SerTeam
