+++ services/serUser.js
@@ -1,26 +1,8 @@
 const ModelUser = require('../models/modUser')
-const ModelTeam = require('../models/modTeam')
-const ModelBelong = require('../models/modBelog')
 
 class serUser {
   constructor () {
     this.modUser = new ModelUser()
-    this.modTeam = new ModelTeam()
-    this.modBelong = new ModelBelong()
-    // this.modMatching = new ModelMatching()
-  }
-
-  // 나의 개별 팀 찾기
-  async findMyTeamList (userId) {
-    try {
-      const teamListOwner = await this.modTeam.findMyTeamList(userId).then()
-      const teamListId = await this.modBelong.findMyTeamList(userId).then()
-      const teamList = teamListOwner.concat(teamListId)
-      // console.log('teamList', teamList);
-      return teamList
-    } catch (error) {
-      console.log(error)
-    }
   }
 
   async findUserInfoById (userId) {
